Tools: Script for building Blender extensions
Utility script that quickly builds extension using Blender's command line arguments. Running `python build` will package the contents of the /source/ and put it outside of the root folder. It also accepts --legacy argument to build it for Blender 4.1 and earlier (with root folder inside the .zip file).

Blender's command line arguments need to run in Blender executable directory, which can be specified with --exe argument. If not provided it defaults to Blender 4.3 installation folder on Windows.